# tsdb: reduce sleep time when reading WAL