--- a/tsdb/head_wal.go
+++ b/tsdb/head_wal.go
@@ -438,7 +438,7 @@ func (wp *walSubsetProcessor) waitUntilIdle() {
 	}
 	wp.input <- []record.RefSample{}
 	for len(wp.input) != 0 {
-		time.Sleep(1 * time.Millisecond)
+		time.Sleep(10 * time.Microsecond)
 		select {
 		case <-wp.output: // Allow output side to drain to avoid deadlock.
 		default: